Corected Task 2_1,Added Task 3_1

+++ Lesson 2/Task 2_2/main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main()
 {
     cout << "Enter your number" << endl;
-    long int number = 0,endNumber = 0;
+    int number = 0,endNumber = 0;
     cin >> number;
     do{
         if ((number % 2) == 0){
